Services - fix index page to join correctly

+++ app/models/Service.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Service(name: String, price: BigDecimal, categoryId: Int, active: Boolean, id: Long = 0)
+case class Service(name: String, price: BigDecimal, categoryId: Long, active: Boolean, id: Long = 0)
